refactor: moved Connector.java & Validator.java inside 'util/' package
